[Gradle] Introduce separate code style configuration for generated code
